+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## UNRELEASED
 
+### Fixed
+* **BREAKING**: Tooltip positioneren (#281)
+
 ## 6.0.1 - 27-03-2019
 
 ### Fixed